% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb2007}
\alias{bb2007}
\title{Coral reef benthic line-intercept transect data from 2007}
\format{}
\source{
\url{http://metadata.imas.utas.edu.au/geonetwork/srv/eng/metadata.show?uuid=2b7b9e49-b426-4a50-a24b-d793e4a58e6e}
}
\usage{
bb2007
}
\description{
Coral reef benthic community transect data from Blue Bowl Reef, Wakatobi Marine National Park, Sulawesi Indonesia.
}
\references{
Haapkyla J, Melbourne-Thomas J, Flavell M, Ward D (2016). Hoga coral reef benthic community transect data.
Centre for Tropical Waters and Aquatic Research, James Cook University. DOI: 10.4226/77/587e953ebfdb3

Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (submitted).
Detecting ecological regime shifts from transect data.
}
\keyword{datasets}
