% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb2010}
\alias{bb2010}
\title{Coral reef benthic line-intercept transect data from 2010}
\format{}
\source{
\url{https://doi.org/10.4226/77/587E953EBFDB3}
}
\usage{
bb2010
}
\description{
Coral reef benthic community transect data from Blue Bowl Reef, Wakatobi Marine National Park, Sulawesi Indonesia.
}
\references{
Haapkyla J, Melbourne-Thomas J, Flavell M, Ward D (2016). Hoga coral reef benthic community transect data.
Centre for Tropical Waters and Aquatic Research, James Cook University. DOI: 10.4226/77/587e953ebfdb3

Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (2018).
Detecting ecological regime shifts from transect data. Ecological Monographs.
}
\keyword{datasets}
