% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CLS1D-package}
\alias{CLS1D-package}
\title{Characteristic Length Scales from 1D Spatial Data.}
\description{
Provides facilities for computing the characteristic length scale
of ecological systems from one dimensional spatial data.
}
\author{
Delphi Ward, Simon Wotherspoon
}
