% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{CLS1D}
\alias{CLS1D}
\title{Sliding window prediction error computation}
\usage{
CLS1D(x, Lhalf, d, metric = c("PRSq", "errX"), n.samples = 10, k = 7,
  replace = TRUE)
}
\arguments{
\item{x}{One-dimensional transect data.}

\item{Lhalf}{The maximum half window size.}

\item{d}{The embedding dimension.}

\item{metric}{Metric to calculate. Options are \code{"errX"} (Error X) or \code{"PRSq"} (prediction r-squared). Error X is the preferred option for this sliding window approach.}

\item{n.samples}{The number of resamples to draw from the embedding. Default is \code{10}.}

\item{k}{The number of neighbours to use in prediction. Default is \code{7}.}

\item{replace}{Sample from the embedding with replacement. Default is \code{TRUE}.}
}
\value{
A matrix of Prediction r-squared or Errror X estimates in which each row is a different window
size and each column is a different subsample of the delay embedding.
}
\description{
Compute the prediction error spectra for a range of window sizes using the sliding window approach,
for Characteristic Length Scale (CLS) estimation.
}
\details{
The transect \code{x} must be numeric with each element corresponding to the presence \code{1}
(or concentration > 0) and absence \code{0} of the target species at one point along the transect.
The spatial resolution must be constant along the length of the transect and optimally at as fine
a resolution as practical.
The density of the target species in windows of size \code{2} to \code{2*Lhalf} is computed with
the \code{\link{window50}} function. Using the \code{\link{embed50}} function an embedding is created for each
window size into \code{d} dimensions. Using the \code{\link{knn}} function the \code{k} nearest neighbours
of each \code{d} dimensional point in the embedding are weighted according to their inverse distance,
and the trajectory to the \code{d+1} dimension predicted. Difference between the predicted and observed
trajectory is calculated either as Error X \code{"errX"}  or Prediction r-squared \code{"PRsq"}.

\deqn{Error X = \sqrt{l\times E_t\times \bigg[\Big(X_l^t - \hat{X_l^t}\Big)^2\bigg]}}

Where \eqn{l} is the window size,  \eqn{X_l^t} is the observed and \eqn{X_l^t} is the predicted
density of species \eqn{X}, and \eqn{E_t} is the expectation of their difference (mean difference).

\deqn{Prediction  r^2 = 1 - \frac{E_t\times \bigg[\Big(X_l^t - \hat{X_l^t}\Big)^2\bigg]}{Var(X_l^t)}}

Note that Error X is the preferred metric for this sliding window approach.
Choice of Lhalf will be constrained by the length of the transect. For example, if \code{d=3}, Lhalf should be
0.49\% of the transect length to have 200 replicates or 0.96\% to obtain 100 replicates of the largest
window size.

The CLS is defined as the window length at which the spectra reaches a plateau (when plotted).
}
\examples{
#Calculate Error X from Blue Bowl reef in 2007
set.seed(2)
x <- CLS1D(ifelse(bb2007 == "CF", 1, 0), Lhalf=50, d=3, metric="errX", n.samples=20)
plotCLS(x, metric="errX")
#The CLS is 60-70 cm

#Compare that to the Error X spectra from the same reef in 2010
set.seed(2)
x <- CLS1D(ifelse(bb2010 == "CF", 1, 0), Lhalf=50, d=3, metric="errX", n.samples=20)
plotCLS(x, metric="errX")
#The CLS has declined from 60-70cm to around 30-40 cm.

#Now check the Prediction r-squared spectra
x <- CLS1D(ifelse(bb2007 == "CF", 1, 0), Lhalf=50, d=3, metric="PRSq", n.samples=20)
plotCLS(x, metric="PRSq")
}
\references{
Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (submitted).
Detecting ecological regime shifts from transect data.
}
\seealso{
\code{\link{window50}},\code{\link{embed50}}, \code{\link{knn}}
}
