% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{CLS1Dsts}
\alias{CLS1Dsts}
\title{Short time-series prediction error computation}
\usage{
CLS1Dsts(x, Lhalf, metric = c("PRSq", "errX"), n.samples = 10, k = 7,
  replace = TRUE)
}
\arguments{
\item{x}{One-dimensional transect data.}

\item{Lhalf}{The maximum half window size.}

\item{metric}{Metric to calculate. Options are \code{"errX"} (Error X) or \code{"PRSq"} (prediction r-squared). Error X is the preferred option for this sliding window approach.}

\item{n.samples}{The number of resamples to draw from the embedding. Default is \code{10}.}

\item{k}{The number of neighbours to use in prediction. Default is \code{7}.}

\item{replace}{Sample from the embedding with replacement. Default is \code{TRUE}.}
}
\value{
A matrix of Prediction r-squared or Errror X estimates in which each row is a different window
size and each column is a different subsample of the delay embedding.
}
\description{
Compute the prediction error spectra for a range of window sizes using the short time-series
approach, for Characteristic Length Scale (CLS) estimation.
}
\details{
Takes a matrix or dataframe \code{x} where each column contains transect data from a different time step,
and each row contains numeric data on the presence \code{1} (or concentration > 0) and absence \code{0}
of the target species at one point along the transect. The spatial resolution must be constant along the
length of the transect and optimally at as fine a resolution as practical.
The density of the target species in each window is computed with the \code{\link{window50sts}} function.
Then, using the \code{\link{knn}} function the \code{k} nearest neighbours of each point in the embedding are
weighted according to their inverse distance, and the trajectory to the final dimension (column) predicted.
Difference between the predicted and observed trajectory is calculated either as Error X \code{"errX"}  or
Prediction r-squared \code{"PRsq"}.

\deqn{Error X = \sqrt{l\times E_t\times \bigg[\Big(X_l^t - \hat{X_l^t}\Big)^2\bigg]}}

Where \eqn{l} is the window size,  \eqn{X_l^t} is the observed and \eqn{X_l^t} is the predicted
density of species \eqn{X}, and \eqn{E_t} is the expectation of their difference (mean difference).

\deqn{Prediction  r^2 = 1 - \frac{E_t\times \bigg[\Big(X_l^t - \hat{X_l^t}\Big)^2\bigg]}{Var(X_l^t)}}
}
\examples{
x <- CLS1Dsts(Sp2R1_sts, Lhalf=20, metric="errX", n.samples=20)
plotCLS(x, metric="errX")

x <- CLS1Dsts(Sp2R2_sts, Lhalf=20, metric="errX", n.samples=20)
plotCLS(x, metric="errX")

x <- CLS1Dsts(Sp2R1_sts, Lhalf=20, metric="PRSq", n.samples=20)
plotCLS(x, metric="PRSq")

x <- CLS1Dsts(Sp2R2_sts, Lhalf=20, metric="PRSq", n.samples=20)
plotCLS(x, metric="PRSq")
}
\references{
Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (submitted).
Detecting ecological regime shifts from transect data.
}
\seealso{
\code{\link{window50sts}}, \code{\link{knn}}
}
