% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{embed50}
\alias{embed50}
\title{Create a Delay Embedding}
\usage{
embed50(x, lhalf, dim)
}
\arguments{
\item{x}{The spatial transect data.}

\item{lhalf}{The window size increment.}

\item{dim}{The embedding dimension + 1.}
}
\value{
An embedding as a matrix.
}
\description{
Form all embeddings of dimension \code{d} for window size \code{2*lhalf} where each window
along the transect overlaps the previous by \code{lhalf}.
}
\details{
Takes one-dimensional spatial (transect) presence-absence or concentration data and creates a
delay embedding of density along the transect. The transect \code{x} must be formatted
so that each element in the vector x represents the occupancy of one species,
morphotype or habitat in one position along the transect. Returns a matrix where
each row is the trajectory of species density along \code{d+1} windows.
}
\examples{
x <- seq(1, 100, by=1)
embed50(x, 5, 4)
}
\seealso{
\code{\link{window50}}
}
