% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{knn}
\alias{knn}
\title{K-Nearest Neighbour Prediction}
\usage{
knn(emb, k = 7)
}
\arguments{
\item{emb}{An embedding constructed with \code{\link{window50}}.}

\item{k}{The number of neighbours.}
}
\value{
Returns a list with two components
\item{\code{obs}}{the observed responses.}
\item{\code{pred}}{the predicted responses.}
}
\description{
Compute k-nearest neighbour predictions for the last dimension of
an embedding.
}
\details{
This function is based on the \code{\link[RANN]{nn2}} function from \pkg{RANN}.
Given an array representing an embedding, it constructs inverse
distance weighted k-nearest neighbour predictions for the last
column of the embedded array.
}
\seealso{
\code{\link{window50}},\code{\link{embed50}},\code{\link[RANN]{nn2}}
}
