% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{plotCLS}
\alias{plotCLS}
\title{Plot prediction error spectra}
\usage{
plotCLS(X1, Lhalf = nrow(X1), metric, ylim, xlab, xticks = 10)
}
\arguments{
\item{X1}{Matrix of prediction error estimates returned by \code{\link{CLS1D}} or \code{\link{CLS1Dsts}}}

\item{Lhalf}{The maximum half window size for which prediction error was estimated. Should be equal to \code{nrow(X1)}.}

\item{metric}{Optional. Specify metric for y-axis label. Options are \code{"errX"} (Error X) or \code{"PRSq"} (prediction r-squared).}

\item{ylim}{Optional. Specify y-axis range. Default is \code{range(X1)}.}

\item{xlab}{Optional. Specify x-axis label. Default is "Window length".}

\item{xticks}{Optional. Specify spacing of ticks on x-axis. Default is by 10.}
}
\value{
Returns a plot of the prediction error spectra.
}
\description{
Plot the prediction error spectra for a range of window sizes, for estimation of the Characteristic
Length Scale (CLS).
}
\details{
Takes matrix of Error X or Prediction r-squared estimates from \code{\link{CLS1D}} or \code{\link{CLS1Dsts}}.
Uses \code{\link[zoo]{rollapply}} from package \pkg{zoo} to estimate the rolling mean (rolling window width = 3)
of the estimates for each window size, and plots with pointwise 95 percent confidence intervals.
}
\examples{
x <- CLS1D(ifelse(bb2007 == "CF", 1, 0), Lhalf=50, d=3, metric="errX", n.samples=20)
plotCLS(x, metric="errX")
}
