% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regime2_sw}
\alias{regime2_sw}
\title{Simulated spatial model output transect data}
\format{}
\usage{
regime2_sw
}
\description{
Transect data from the second regime of Test Case 2 in Ward et al (submitted)
}
\references{
Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (2018).
Detecting ecological regime shifts from transect data. Ecological Monographs.
}
\keyword{datasets}
