% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sp2R2_sts}
\alias{Sp2R2_sts}
\title{Simulated model short time-series transect data}
\format{}
\usage{
Sp2R2_sts
}
\description{
Transect data from three time steps (separated by 10 time steps) of the second regime of Test Case 2 in Ward et al (submitted).
This example contains only Species 2 presence-absence data.
}
\references{
Ward D, Wotherspoon S, Melbourne-Thomas J, Haapkyla J, Johnson CR (2018).
Detecting ecological regime shifts from transect data. Ecological Monographs.
}
\keyword{datasets}
