% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{window50}
\alias{window50}
\title{Compute averages in sliding window}
\usage{
window50(x, lhalf)
}
\arguments{
\item{x}{One-dimensional spatial presence absence data.}

\item{lhalf}{The window size increment.}
}
\value{
A numeric vector containing mean density in each window of size \code{2*lhalf}.
}
\description{
Compute averages in windows of length \code{2*lhalf} along a transect,
where each window overlaps the previous window by lhalf.
}
\details{
Computes averages in sliding windows along transect by taking differences of
cumulative sums.
}
