% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls1d.r
\name{window50sts}
\alias{window50sts}
\title{Compute averages for short time-series}
\usage{
window50sts(x, lhalf)
}
\arguments{
\item{x}{Data frame or matrix of 1-dimensional spatial transects}

\item{lhalf}{the window size increment.}
}
\value{
a matrix or dataframe containing mean density in each window of size \code{2*lhalf}
}
\description{
Compute averages in windows of length \code{2*lhalf} along each transect in a set of time-delayed
transects. Each window overlaps the previous window by lhalf within the same transect. Between transects
the windows are in the same position spatially.
}
\details{
Computes averages in sliding windows along transects by taking differences of
cumulative sums.
}
